Port Pantheist code (again)
Originally, Bakko was intended to start out as a Kotlin version
of the Pantheist (decompiled as I lost the code). It turned out that
either I screwed up the port badly, or that Kotlin just doesn't
perform that well as a Robocode language.

This commit finishes taking over Pantheist's code to Bakko. What's
left is refactoring the mess and separating the different parts.