Circular prediction gun and weapon management
I added Bakko's circular prediction gun (from the Kotlin version).
It is not in use yet, but I also introduced a weapon manager class
responsible for handling the events of multiple weapons.

The next step is to calculate the (virtual) hit rate of the
gun(s) to figure out which is best to use.

Further in the future, I could persist the information which gun
was best against which bot.